+++ changelog.md
@@ -1,7 +1,7 @@
 This release brings the following changes:
 
-* Handle server connections which break when client sends ALPN information. This should restore connectivity to nvdaremote.com and similar servers.
+* Add missing libraries required in older NVDA releases.
 
 Important: some anti-virus software may fla parts of this add-on as malicious. Specifically, `url_handler.exe`, which opens `remote://` and `tele://` links. If you don't use this feature, you can safely quarantine or delete the file. Otherwise, you must add it as an exception.
 
-SHA256:
+SHA256: